+++ src/Graphics/Model/mdlManager.c
@@ -1226,7 +1226,7 @@ void func_0031b470(void);
 void func_0031b4a0(char* param_1,u16 param_2);
 u32 func_0031b680(int param_1,int param_2,int *param_3,int *param_4);
 void func_0031b820(u64 param_1,u64 param_2);
-u32 func_0031be80(u64 param_1);
+u32 func_0031be80(u32 param_1);
 void func_0031c000(char* param_1,u16 param_2);
 u32 func_0031c1d0(int param_1);
 u32 func_0031c7e0(int param_1);
@@ -9080,7 +9080,7 @@ void func_0031b820(u64 param_1,u64 param_2)
 // FUN_0031BE80 NONMATCHING
 
 
-u32 func_0031be80(u64 param_1)
+u32 func_0031be80(u32 param_1)
